Introduce quarry with lvm backend
Quarry is a front-end for storage provisioning and management.  The
market is filled with different SAN and NAS products each with its own
native management APIs.  By using quarry, an application will gain
support for all participating storage vendors while still taking
advantage of the benefits of the individual array being used (eg. array
offloading).

To begin the API design we introduce an LVM backend which exposes LVs in
a special VG as ISCSI LUNs.  The initial operations available are:
 - Create volume/snapshot
 - Delete volume/snapshot
 - Create a volume connection
 - Delete a volume connection

Signed-off-by: Adam Litke <alitke@redhat.com>